further abstracted jetty-specific stuff into dedicated interface
